Configure /etc/hostname in clones
This commit updates the clone type to populate /etc/hostname in the uppderdir
of the clone's overlay to contain a FQDN formed by concatenating the clone's
name with the host system's certname. This change helps the puppet
service running on the clones pick to a certname that is distinct from that
used by the host without needing to set certname in puppet.conf.

Avoiding modifications to puppet.conf preserves the ability to
coordinate mass updates of puppet.conf by editing the base image
in the lowerdir of the overlay.