Swap Puppet mount resource for SystemD mount units
This commit updates the clone type to create SystemD mount units instead
of using Puppet's mount type to manage the overlay file systems used by
the clone. This change results in the filesystems being automatically
mounted when clones start and unmounted when they stop. Re-starting a
clone results in the filesystem being remounted, which is useful for
ensuring modifications made to the upper or lower layers of the overlay
are refreshed.

A `puppet-clone-army.target` is also added that can be used to start or
stop all clones en masse.