Don't manage running state of individual clones
This commit updates the Clone type to skip enforcing a "running" or
"stopped" state on the clones. This prevents the puppet agent on
the host from interrupting experiments by doing a mass start or stop
of clones at a bad time.

Instead, clones are simply enabled so their `[Install]` configuration
attaches to the `puppet-clone-army.target` unit. This allows mass
starts, stops, or restarts to be coordinated through the target.