Re-order Base_image resources
This commit re-orders the resources in the Base_image type to place
puppet-agent installation last and adds some comments marking different
sections. This fixes an issue where puppet-agent installation would
fail on the first catalog application when /etc/hosts and /etc/resolv.conf
need to be configured in order for the master's hostname to be resolved.